Ship the nativescript CLI monkeypatch as a published binary, not a demo-app script
ember-native is a library other apps install, so a fix that only lives in
demo-app/scripts/ never reaches consumers hitting the same upstream
`nativescript` CLI bug. Checked for an upstream fix first: none released yet
(latest stable is still 9.0.6), though a 9.0.7-next/9.1.0 prerelease already
has it, confirmed by reading its compileWithoutWatch source directly.

Moves the monkeypatch into ember-native/utils/nativescript-cli.js, published
via a new `ember-native-nativescript` bin entry so any consuming app can use
it directly. Resolves the nativescript dependency from the invoking project's
cwd rather than the addon's own install location (which never depends on
nativescript itself), and feature-detects the upstream fix so the binary
becomes a transparent passthrough once a consumer upgrades past it.
