+++ demo-app/package.json
@@ -5,12 +5,12 @@
   "private": true,
   "scripts": {
     "devtools": "ember-devtools",
-    "run": "node scripts/nativescript-cli.js debug android",
-    "test": "node scripts/nativescript-cli.js test android --config nativescript.test.vite.config.ts --no-watch",
-    "build": "node scripts/nativescript-cli.js build android",
-    "debug-test": "node scripts/nativescript-cli.js test android --debug-brk --config nativescript.test.vite.config.ts --no-watch",
-    "debug": "node scripts/nativescript-cli.js debug android --debug-brk",
-    "prepare-android": "node scripts/nativescript-cli.js prepare android",
+    "run": "ember-native-nativescript debug android",
+    "test": "ember-native-nativescript test android --config nativescript.test.vite.config.ts --no-watch",
+    "build": "ember-native-nativescript build android",
+    "debug-test": "ember-native-nativescript test android --debug-brk --config nativescript.test.vite.config.ts --no-watch",
+    "debug": "ember-native-nativescript debug android --debug-brk",
+    "prepare-android": "ember-native-nativescript prepare android",
     "lint": "concurrently 'pnpm:lint:*(!fix)' --names 'lint:'",
     "lint:fix": "concurrently 'pnpm:lint:*:fix' --names 'fix:'",
     "lint:hbs": "ember-template-lint . --no-error-on-unmatched-pattern",
